Don't re-use cached base fields when TZ changed and using :localtime timezone
When /etc/localtime is missing, test clock-38.1 fails on Linux. This is a result of re-using
cached dataPtr->lastBase.date when the timezone and lastBase timezone are both :localtime.
Unfortunately, this can't be done safely with :localtime unless we know that the old base
had was computed when TZ had the same value.